Create full def for unzip
